@@ -1,4 +1,8 @@
 export interface SearchCriteria {
   query?: string;
-  userId?: string;
+  userId?: number;
+  branchId?: number;
+  admin?: boolean;
+  simId?: number;
+  transactionType?: string  ;
 }
