@@ -3,10 +3,11 @@ import { isPlatformBrowser, CommonModule } from '@angular/common';
 import { CardModule } from 'primeng/card';
 import { ChartModule } from 'primeng/chart';
 import { RouterModule } from '@angular/router';
+import { HeadCardComponent } from "../common/head-card/head-card.component";
 
 @Component({
     selector: 'app-dashboard',
-    imports: [CommonModule, CardModule, ChartModule, RouterModule],
+    imports: [CommonModule, CardModule, ChartModule, RouterModule, HeadCardComponent],
     templateUrl: './dashboard.html',
     styleUrls: ['./dashboard.scss']
 })
@@ -15,6 +16,13 @@ export class Dashboard {
   lineData: any;
   doughnutData: any;
   chartOptions: any;
+
+  statistics = [
+    { label: 'إجمالي الرصيد', value: 5750.00 , icon: 'pi pi-wallet', color: '#1B0F31' },
+    { label: 'الرصيد الرئيسي', value:  3450.00, icon: 'pi pi-dollar', color: '#d4211eff' },
+    { label: 'الرصيد الائتماني', value: 2300.00, icon: 'pi pi-credit-card', color: '#0C4842' },
+    { label: 'إجمالي التحملات', value: 1500, icon: 'pi pi-chart-line', color: '#07403C' }
+  ];
 
   constructor() {
     this.isBrowser = isPlatformBrowser(inject(PLATFORM_ID));
@@ -29,7 +37,7 @@ export class Dashboard {
           data: [1000, 1600, 1400, 1800, 2300],
           fill: true,
           tension: 0.4,
-          borderColor: '#1B0F31',
+          borderColor: '#4D20D2',
           backgroundColor: 'rgba(12, 72, 66, 0.1)'
         }
       ]
@@ -40,7 +48,7 @@ export class Dashboard {
       datasets: [
         {
           data: [45, 30, 25],
-          backgroundColor: ['#1B0F31', '#d4211eff', 'rgba(12, 72, 66, 0.1)'],
+          backgroundColor: ['#4D20D2', '#d4211eff', 'rgba(12, 72, 66, 0.1)'],
           borderWidth: 1
         }
       ]
