@@ -11,7 +11,7 @@ import { environment } from '../../environments/environment';
 export class UserService {
   private http = inject(HttpClient);
 
-  listUsers(searchCriteria: SearchCriteria, page: number, size: number): Observable<User[]> {
+  listUsers(searchCriteria: SearchCriteria = {}, page: number=0, size: number = 1000): Observable<User[]> {
     return this.http.post<User[]>(`${environment.expressCashApiUrl}/users?page=${page}&size=${size}`, searchCriteria);
   }
 }
